fix(backend): fix AuthControllerTest and refactor MyBatisPlus config
- Moved @MapperScan to MyBatisPlusConfig to prevent slice tests from loading full MyBatis context
- Updated AuthControllerTest with SecurityConfig import and filter mock passthrough
- Removed deprecated user tag services/controllers
- Cleaned up pom.xml and updated AGENTS.md

+++ backend/src/main/java/com/douyin/entity/VideoStatsDaily.java
@@ -17,15 +17,15 @@ public class VideoStatsDaily {
 
     private LocalDate date;
 
-    private Integer imprCnt;
+    private Long imprCnt;
 
-    private Integer clickCnt;
+    private Long clickCnt;
 
-    private Integer likeCnt;
+    private Long likeCnt;
 
-    private Integer finishCnt;
+    private Long finishCnt;
 
-    private Integer shareCnt;
+    private Long shareCnt;
 
     /**
      * Total watch time in milliseconds.
